+++ projects/doc-web-site/src/shared/index.ts
@@ -1,7 +1,0 @@
-export * from './shared.module';
-
-export * from './models/menu';
-export * from './models/menu-factory';
-export * from './models/version-factory';
-export * from './constants/PATH';
-export * from './constants/VERSION';