+++ projects/test-web-site/src/app/app.component.ts
@@ -1,17 +1,15 @@
 import {
-  ChangeDetectionStrategy, ChangeDetectorRef, Component, OnDestroy, OnInit
+  ChangeDetectionStrategy, ChangeDetectorRef, Component, OnInit
 } from '@angular/core';
-import {
-  TranslationChange, TranslatableLanguageList, TranslatableOptionList,
-  TranslationService
-} from 'ng-translation';
+import { TranslationChange, TranslationService } from 'ng-translation';
+import { TranslatableLanguageList, TranslatableOptionList } from 'ng-translatable';
 
-@Component({
+@Component( {
   selector: 'app-root',
   templateUrl: './app.component.html',
-  styleUrls: ['./app.component.css'],
+  styleUrls: [ './app.component.css' ],
   changeDetection: ChangeDetectionStrategy.OnPush
-})
+} )
 export class AppComponent implements OnInit {
 
   initialized = false;
@@ -19,7 +17,7 @@ export class AppComponent implements OnInit {
   languages: TranslatableLanguageList;
 
   constructor(
-    private  cdRef: ChangeDetectorRef,
+    private cdRef: ChangeDetectorRef,
     private translate: TranslationService
   ) {
     this.translate.statusChange.subscribe( this.ngtChanges.bind( this ) );