+++ projects/test-web-site/src/app/winter/index/index.component.ts
@@ -1,9 +1,0 @@
-import { ChangeDetectionStrategy, Component } from '@angular/core';
-
-@Component({
-  selector: 'app-index',
-  templateUrl: './index.component.html',
-  styleUrls: ['./index.component.css'],
-  // changeDetection: ChangeDetectionStrategy.OnPush
-})
-export class IndexComponent { }