+++ projects/test-web-site/src/tests/null/null.component.ts
@@ -1,0 +1,12 @@
+import { Component, ChangeDetectionStrategy } from '@angular/core';
+
+@Component( {
+  selector: 'app-null',
+  templateUrl: './null.component.html',
+  styleUrls: [ './null.component.css' ],
+  changeDetection: ChangeDetectionStrategy.OnPush
+} )
+export class NullComponent {
+
+  today = new Date( Date.now() );
+}